Error Handling in API calls

+++ src/Components/Menu/Menu.jsx
@@ -3,6 +3,7 @@ import { Sidebar, Segment, Header, Button } from 'semantic-ui-react'
 import { connect } from 'react-redux'
 import { withRouter } from 'react-router-dom'
 import { Link } from 'react-router-dom'
+import { wipeCredentials } from '../../Modules/wipeCredentials'
 import axios from 'axios'
 import i18n from '../../i18n'
 import { withTranslation } from 'react-i18next'
@@ -16,24 +17,23 @@ class Menu extends Component {
   signOut = (e) => {
     e.preventDefault()
     const { t } = this.props
-    const path = '/api/v1/auth/sign_out'
-    const headers = {
-      uid: window.localStorage.getItem('uid'),
-      client: window.localStorage.getItem('client'),
-      'access-token': window.localStorage.getItem('access-token')
+    if (window.navigator.onLine === false) {
+      window.alert(t('reusable:errors:window-navigator'))
+    } else {
+      const path = '/api/v1/auth/sign_out'
+      const headers = {
+        uid: window.localStorage.getItem('uid'),
+        client: window.localStorage.getItem('client'),
+        'access-token': window.localStorage.getItem('access-token')
+      }
+      axios.delete(path, { headers: headers })
+        .then(() => {
+          wipeCredentials('/')
+        })
+        .catch(() => {
+          wipeCredentials('/login')
+        })
     }
-    axios.delete(path, { headers: headers })
-      .then(() => {
-        window.localStorage.removeItem('access-token')
-        window.localStorage.removeItem('token-type')
-        window.localStorage.removeItem('client')
-        window.localStorage.removeItem('uid')
-        window.localStorage.removeItem('expiry')
-        window.location.replace('/')
-      })
-      .catch(() => {
-        window.alert(t('reusable:errors:sign-out-error'))
-      })
   }
 
   changeLng(lng) {
